♻️: 食品データのパース処理を共通化する
